refactor: 💡 checkout page
