+++ domains/core/components/ui/OrderSummary.vue
@@ -11,20 +11,20 @@ const { cart } = useCart();
       class="flex justify-between items-end py-2 px-4 md:px-6 md:pt-6 md:pb-4"
     >
       <p class="typography-headline-4 font-bold md:typography-headline-3">
-        {{ $t('orderSummary') }}
+        {{ $t("orderSummary") }}
       </p>
       <p class="typography-text-base font-medium" data-testid="total-in-cart">
         {{
-          $t('itemsInCart', { count: cart?.order?.websiteOrderLine?.length })
+          $t("itemsInCart", { count: cart?.order?.websiteOrderLine?.length })
         }}
       </p>
     </div>
     <div class="px-4 pb-4 mt-3 md:px-6 md:pb-6 md:mt-0">
       <div class="flex justify-between typography-text-base mb-2">
         <div class="flex flex-col grow pr-2 gap-2">
-          <p>{{ $t('itemsSubtotal') }}</p>
-          <p>{{ $t('delivery') }}</p>
-          <p>{{ $t('estimatedTax') }}</p>
+          <p>{{ $t("itemsSubtotal") }}</p>
+          <p>{{ $t("delivery") }}</p>
+          <p>{{ $t("estimatedTax") }}</p>
         </div>
         <div class="flex flex-col gap-2 text-right">
           <p data-testid="special-price">
@@ -41,7 +41,7 @@ const { cart } = useCart();
         class="flex justify-between typography-text-base mb-2"
       >
         <p class="flex grow pr-2">
-          {{ $t('discounts', { count: cart.order.coupons.length }) }}
+          {{ $t("discounts", { count: cart.order.coupons.length }) }}
         </p>
         <p class="flex text-right">
           {{ $currency(Number(cart.order?.amountDiscounts)) }}
@@ -52,7 +52,7 @@ const { cart } = useCart();
         class="flex justify-between typography-text-base mb-4"
       >
         <p class="flex grow pr-2">
-          {{ $t('giftCard', { count: cart.order.giftCards.length }) }}
+          {{ $t("giftCard", { count: cart.order.giftCards.length }) }}
         </p>
         <p class="flex text-right">
           {{ $currency(Number(cart.order?.amountGiftCards)) }}
@@ -61,7 +61,7 @@ const { cart } = useCart();
       <div
         class="flex justify-between typography-headline-4 md:typography-headline-3 font-bold pb-4 mb-4"
       >
-        <p>{{ $t('total') }}</p>
+        <p>{{ $t("total") }}</p>
         <p data-testid="total">
           {{ $currency(Number(cart?.order?.amountTotal)) }}
         </p>
